Correction of sentence in open source tab
Developers can access and leverage on the various technologies and tools that Government agencies have open-sourced. This includes COVID-19 open-source technologies such as TraceTogether and VigilantGantry.

To 

Developers can access and leverage the various technologies and tools that Government agencies have open-sourced. This includes COVID-19 open-source technologies such as TraceTogether and VigilantGantry.